requirements.txt
matplotlib==3.1.2
numpy==1.17.4
PyAutoGUI==0.9.48
scikit_image==0.16.2
Pillow==6.2.1
skimage==0.0
